Add creation of lock file in master process
Previously, the lock was removed in the master process,
while some file lock implementations check for the file existence
and and create it themselves, for example:
https://github.com/tox-dev/filelock/blob/19f78fcdb30c8ce296acd54a016a1247e27dac3e/src/filelock/_unix.py#L40

Looks like this may lead to race conditions (faced this after update of
docker on Mac).

This change ensures that lock file exists before worker processes starts
to collect tests.