Update xfailed tests handling
In the previous implementation, tests marked as
`xfailed` were displayed in qase.io as `skipped`.
But this is actually incorrect.

`xfail` implies that the test failed for some already
known reason. So we use the `blocked` status so
as not to mislead the QA team. The `blocked`
status will let them know that the case is blocked
for some reason described in the `xfail` comment
(this comment will be duplicated as a result of the
case execution).